requirements.txt
numpy
opencv-python
pytesseract
matplotlib
Pillow
imutils
pyspellchecker

# pip install -r requirements.txt